create kafka route component
